+++ frontend/app/(tabs)/_layout.tsx
@@ -11,8 +11,10 @@
  */
 import * as React from "react";
 import { Tabs } from "expo-router";
+import { Ionicons } from "@expo/vector-icons";
 import { useSafeAreaInsets } from "react-native-safe-area-context";
 import { AuthGuard } from "../../components/AuthGuard";
+import { colors } from "../../theme/colors";
 
 export default function TabsLayout() {
   const insets = useSafeAreaInsets();
@@ -25,6 +27,8 @@ export default function TabsLayout() {
       {
         screenOptions: {
           headerShown: false,
+          tabBarActiveTintColor: colors.primary,
+          tabBarInactiveTintColor: colors.textSecondary,
           tabBarStyle: {
             margin: 0,
             padding: 0,
@@ -35,13 +39,37 @@ export default function TabsLayout() {
         },
       },
     // Home feed / dashboard tab - main content feed
-    React.createElement(Tabs.Screen, { name: "home" }),
+    React.createElement(Tabs.Screen, {
+      name: "home",
+      options: {
+        title: "Home",
+        tabBarIcon: ({ color, size }) =>
+          React.createElement(Ionicons, { name: "home", color, size }),
+      },
+    }),
     // App settings and preferences tab - app configuration and preferences
-    React.createElement(Tabs.Screen, { name: "profile" }),
-    // Navigation tab - camera frame capture and transmission
-    React.createElement(Tabs.Screen, { name: "nav-dev" }),
-    // Sensor dev tab - hidden from tab bar (dev only, accessible via direct navigation)
-    React.createElement(Tabs.Screen, { name: "sensor-dev" }),
+    React.createElement(Tabs.Screen, {
+      name: "profile",
+      options: {
+        title: "Profile",
+        tabBarIcon: ({ color, size }) =>
+          React.createElement(Ionicons, { name: "person", color, size }),
+      },
+    }),
+    // Navigation dev route - hidden from tab bar but still routable
+    React.createElement(Tabs.Screen, {
+      name: "nav-dev",
+      options: {
+        href: null,
+      },
+    }),
+    // Sensor dev route - hidden from tab bar but still routable
+    React.createElement(Tabs.Screen, {
+      name: "sensor-dev",
+      options: {
+        href: null,
+      },
+    }),
     ),
   );
 }
