+++ frontend/app/(tabs)/profile/index.tsx
@@ -8,15 +8,20 @@
  * Uses React.createElement (non-JSX) to match the project's TypeScript configuration.
  */
 import * as React from "react";
-import * as Speech from "expo-speech";
-import { View, Text, Switch, ScrollView } from "react-native";
+import { View, Text, Switch, ScrollView, Alert, ActivityIndicator } from "react-native";
 import { useRouter } from "expo-router";
 import { SafeAreaView } from "react-native-safe-area-context";
 import Button from "../../../components/Button";
 import { useAuth } from "../../../contexts/AuthContext";
-import { getIdToken } from "../../../services/tokenStorage";
+import { canUseBiometrics, promptBiometricUnlock } from "../../../services/biometricAuth";
+import {
+  clearBiometricLoginPreference,
+  getBiometricLoginEnabled,
+  getIdToken,
+  setBiometricLoginEnabled,
+} from "../../../services/tokenStorage";
 import { spacing } from "../../../theme/spacing";
-import { typography } from "../../../theme/typography";
+import { fontFamily, typography } from "../../../theme/typography";
 import { colors } from "../../../theme/colors";
 import { useSettings } from "../../../contexts/SettingsContext";
 
@@ -31,6 +36,9 @@ export default function Profile() {
 
   // App settings state
   const { cameraMode, setCameraMode } = useSettings();
+  const [isBiometricLoginEnabled, setIsBiometricLoginEnabled] = React.useState(false);
+  const [isLoadingBiometricPreference, setIsLoadingBiometricPreference] = React.useState(true);
+  const [isSavingBiometricPreference, setIsSavingBiometricPreference] = React.useState(false);
 
   React.useEffect(() => {
     const fetchUserData = async () => {
@@ -68,6 +76,21 @@ export default function Profile() {
     fetchUserData();
   }, [isDevBypass]);
 
+  React.useEffect(() => {
+    const loadBiometricPreference = async () => {
+      try {
+        const enabled = await getBiometricLoginEnabled();
+        setIsBiometricLoginEnabled(enabled);
+      } catch (error) {
+        console.error("Failed to load biometric preference:", error);
+      } finally {
+        setIsLoadingBiometricPreference(false);
+      }
+    };
+
+    void loadBiometricPreference();
+  }, []);
+
   const handleLogout = async () => {
     try {
       setIsLoggingOut(true);
@@ -85,11 +108,77 @@ export default function Profile() {
     }
   };
 
+  const handleBiometricToggle = async (nextValue: boolean) => {
+    if (isSavingBiometricPreference || isLoadingBiometricPreference) {
+      return;
+    }
+
+    try {
+      setIsSavingBiometricPreference(true);
+
+      if (!nextValue) {
+        await clearBiometricLoginPreference();
+        setIsBiometricLoginEnabled(false);
+        return;
+      }
+
+      const availability = await canUseBiometrics();
+      if (!availability.available) {
+        if (availability.status === "not-supported") {
+          Alert.alert(
+            "Biometrics unavailable",
+            "This device does not support biometric authentication."
+          );
+        } else if (availability.status === "not-enrolled") {
+          Alert.alert(
+            "Biometrics not set up",
+            "Set up Face ID or fingerprint in your device settings and try again."
+          );
+        } else {
+          Alert.alert(
+            "Biometrics unavailable",
+            "We could not verify biometric availability right now. Please try again."
+          );
+        }
+        return;
+      }
+
+      const promptResult = await promptBiometricUnlock("Confirm biometric login setup");
+      if (!promptResult.success) {
+        if (promptResult.status === "cancelled") {
+          Alert.alert("Setup cancelled", "Biometric login was not enabled.");
+        } else if (promptResult.status === "lockout") {
+          Alert.alert(
+            "Biometrics temporarily locked",
+            "Too many attempts were made. Unlock your device and try again."
+          );
+        } else {
+          Alert.alert(
+            "Verification failed",
+            "Biometric verification did not succeed. Biometric login was not enabled."
+          );
+        }
+        return;
+      }
+
+      await setBiometricLoginEnabled(true);
+      setIsBiometricLoginEnabled(true);
+    } catch (error) {
+      console.error("Failed to update biometric login preference:", error);
+      Alert.alert(
+        "Unable to update setting",
+        "We could not update your biometric login preference. Please try again."
+      );
+    } finally {
+      setIsSavingBiometricPreference(false);
+    }
+  };
+
   return React.createElement(
     SafeAreaView,
     {
       style: { flex: 1, backgroundColor: "#FFFFFF" },
-      edges: ["top", "bottom"],
+      edges: ["top"],
     },
     React.createElement(
       ScrollView,
@@ -136,52 +225,188 @@ export default function Profile() {
           View,
           {
             style: {
-              flexDirection: "row",
-              justifyContent: "space-between",
-              alignItems: "center",
+              borderWidth: 1,
+              borderColor: colors.backgroundSecondary,
+              borderRadius: 12,
+              padding: spacing.md,
+              gap: spacing.xs,
             },
           },
           React.createElement(
             View,
-            { style: { flex: 1, paddingRight: spacing.md, gap: spacing.xs } },
-            React.createElement(Text, { style: typography.h3 }, "Camera Mode"),
+            {
+              style: {
+                flexDirection: "row",
+                justifyContent: "space-between",
+                alignItems: "center",
+                gap: spacing.md,
+              },
+            },
+            React.createElement(
+              View,
+              { style: { flex: 1, gap: spacing.xs } },
+              React.createElement(
+                Text,
+                {
+                  style: {
+                    ...typography.body,
+                    fontFamily: fontFamily.bold,
+                    fontWeight: "700",
+                  },
+                },
+                "Camera Mode"
+              ),
+              React.createElement(
+                Text,
+                { style: { ...typography.body, color: colors.textSecondary } },
+                "When enabled, collision detection and live navigation routing are active."
+              )
+            ),
+            React.createElement(Switch, {
+              value: cameraMode,
+              onValueChange: handleCameraModeToggle,
+              thumbColor: cameraMode ? colors.primary : undefined,
+              trackColor: {
+                false: colors.secondary,
+                true: "#98caa8",
+              },
+              accessibilityLabel: "Camera Mode",
+              accessibilityRole: "switch",
+              accessibilityHint: "Enable or disable camera mode",
+            })
+          )
+        ),
+        React.createElement(
+          View,
+          {
+            style: {
+              borderWidth: 1,
+              borderColor: colors.backgroundSecondary,
+              borderRadius: 12,
+              padding: spacing.md,
+              gap: spacing.xs,
+            },
+          },
+          React.createElement(
+            View,
+            {
+              style: {
+                flexDirection: "row",
+                justifyContent: "space-between",
+                alignItems: "center",
+                gap: spacing.md,
+              },
+            },
+            React.createElement(
+              View,
+              {
+                style: {
+                  flex: 1,
+                  gap: spacing.xs,
+                },
+              },
+              React.createElement(
+                Text,
+                {
+                  style: {
+                    ...typography.body,
+                    fontFamily: fontFamily.bold,
+                    fontWeight: "700",
+                  },
+                },
+                "Biometric Login"
+              ),
+              React.createElement(
+                Text,
+                {
+                  style: {
+                    ...typography.body,
+                    color: colors.textSecondary,
+                  },
+                },
+                "Use Face ID or fingerprint to unlock your account on app launch."
+              )
+            ),
+            isLoadingBiometricPreference
+              ? React.createElement(ActivityIndicator, {
+                  size: "small",
+                  color: colors.primary,
+                })
+              : React.createElement(Switch, {
+                  value: isBiometricLoginEnabled,
+                  onValueChange: handleBiometricToggle,
+                  disabled: isSavingBiometricPreference,
+                  thumbColor: isBiometricLoginEnabled ? colors.primary : undefined,
+                  trackColor: {
+                    false: colors.secondary,
+                    true: "#98caa8",
+                  },
+                  accessibilityLabel: "Biometric Login",
+                  accessibilityRole: "switch",
+                  accessibilityHint: "Enable or disable biometric login unlock",
+                })
+          ),
+          isSavingBiometricPreference &&
             React.createElement(
               Text,
-              { 
-                style: { ...typography.body, color: colors.textSecondary },
-                // Make the text act as a button that triggers Text-To-Speech
-                onPress: () => {
-                  // Optional: stop any current speech before reading the new one
-                  Speech.stop(); 
-                  Speech.speak("Camera mode: When enabled, collision detection and live navigation routing are active.");
+              {
+                style: {
+                  ...typography.label,
+                  color: colors.textSecondary,
                 },
-                accessibilityRole: "button",
-                accessibilityLabel: "Read camera mode description",
-                accessibilityHint: "Double tap to hear the description of camera mode aloud",
               },
-              "When enabled, collision detection and live navigation routing are active."
+              "Updating biometric preference..."
             )
-          ),
-          React.createElement(Switch, {
-            value: cameraMode,
-            onValueChange: handleCameraModeToggle,
-            trackColor: { false: "#E5E5E5", true: colors.primary },
-            ios_backgroundColor: "#E5E5E5",
-          })
         ),
 
         // --- Actions Section (Change Password) ---
         React.createElement(
           View,
           { style: { gap: spacing.xl } },
-          React.createElement(Button, {
-            onPress: () => router.push("/profile/change-password" as any),
-            title: "Change Password",
-            variant: "secondary",
-            accessibilityLabel: "Change Password",
-            accessibilityRole: "button",
-            accessibilityHint: "Go to the change password screen",
-          })
+          React.createElement(
+            View,
+            {
+              style: {
+                borderWidth: 1,
+                borderColor: colors.backgroundSecondary,
+                borderRadius: 12,
+                padding: spacing.md,
+                gap: spacing.sm,
+              },
+            },
+            React.createElement(
+              Text,
+              {
+                style: {
+                  ...typography.body,
+                  fontFamily: fontFamily.bold,
+                  fontWeight: "700",
+                },
+              },
+              "Change Password"
+            ),
+            React.createElement(
+              Text,
+              {
+                style: {
+                  ...typography.body,
+                  color: colors.textSecondary,
+                },
+              },
+              "Update your account password to keep your login secure."
+            ),
+            React.createElement(Button, {
+              onPress: () => router.push("/profile/change-password" as any),
+              title: "Change Password",
+              variant: "secondary",
+              accessibilityLabel: "Change Password",
+              accessibilityRole: "button",
+              accessibilityHint: "Go to the change password screen",
+              style: {
+                marginTop: spacing.md,
+              },
+            })
+          )
         )
       ), // End of Top Content Group
 
@@ -190,16 +415,52 @@ export default function Profile() {
       React.createElement(
         View,
         { style: { marginTop: "auto", paddingTop: spacing.xl } },
-        React.createElement(Button, {
-          onPress: handleLogout,
-          title: "Logout",
-          variant: "danger",
-          loading: isLoggingOut,
-          disabled: isLoggingOut,
-          accessibilityLabel: "Logout",
-          accessibilityRole: "button",
-          accessibilityHint: "Sign out of your account",
-        })
+        React.createElement(
+          View,
+          {
+            style: {
+              borderWidth: 1,
+              borderColor: colors.backgroundSecondary,
+              borderRadius: 12,
+              padding: spacing.md,
+              gap: spacing.sm,
+            },
+          },
+          React.createElement(
+            Text,
+            {
+              style: {
+                ...typography.body,
+                fontFamily: fontFamily.bold,
+                fontWeight: "700",
+              },
+            },
+            "Logout"
+          ),
+          React.createElement(
+            Text,
+            {
+              style: {
+                ...typography.body,
+                color: colors.textSecondary,
+              },
+            },
+            "Sign out of your account on this device."
+          ),
+          React.createElement(Button, {
+            onPress: handleLogout,
+            title: "Logout",
+            variant: "danger",
+            loading: isLoggingOut,
+            disabled: isLoggingOut,
+            accessibilityLabel: "Logout",
+            accessibilityRole: "button",
+            accessibilityHint: "Sign out of your account",
+            style: {
+              marginTop: spacing.md,
+            },
+          })
+        )
       )
     )
   );
